gui/requirements.txt
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pillow==10.3.0
requests==2.32.3
vpython==7.6.5
Flask==3.0.3
Flask-MySQLdb==2.0.0
mysqlclient==2.2.4